refactor: 프로필페이지내에 있던 로그아웃 기능을 TopNavBasic 으로 이동 1. 프로필페이지 내에 작성된 로그아웃 기능을 TopNavBasic으로 이동하였습니다.
